Attempt to make docker start the front-end
NOTE: probably broken
I need to transfer this between computers

+++ web/initdb.js
@@ -21,21 +21,31 @@ function readSync(question, silent=false) {
 }
 
 async function run() {
-    await User.deleteMany().catch((err) => console.log(err));
+    // Check if database exists
+    var qsets = QuestionSet.find().catch((err) => console.log(err));
+    if (qsets.length != 0) {
+        console.log("Database already exists, not initialising")
+        process.exit(0)
+    }
+    /*await User.deleteMany().catch((err) => console.log(err));
     await QuestionSet.deleteMany().catch((err) => console.log(err));
-    await Form.deleteMany().catch((err) => console.log(err));
+    await Form.deleteMany().catch((err) => console.log(err));*/
 
-    var number = await readSync('Enter IT number: ');
+    /*var number = await readSync('Enter IT number: ');
     var fname = await readSync('Enter IT first name: ');
-    var lname = await readSync('Enter IT last name: ');
-    var password, password2;
-    password = await readSync('Enter IT password (will not echo): ', slient=true);
+    var lname = await readSync('Enter IT last name: ');*/
+    var number = "00000000"
+    var fname = "Admin"
+    var lname = "Admin"
+    var password = process.env.IT_PASSWORD
+    //var password, password2;
+    /*password = await readSync('Enter IT password (will not echo): ', slient=true);
     password2 = await readSync('Enter IT password again: ', slient=true);
     while (password !== password2) {
         console.log("Passwords do not match");
         password = await readSync('Enter IT password (will not echo): ', slient=true);
         password2 = await readSync('Enter IT password again: ', slient=true);
-    }
+    }*/
     var user = User.create(fname, lname, number)
     user.isIT = true;
     await User.register(user, password).catch((err) => console.log(err));
